Added log_metrics to keyword example and fixed bugs
