+++ fedn/network/grpc/fedn_pb2_grpc.py
@@ -5,7 +5,7 @@ import warnings
 
 from fedn.network.grpc import fedn_pb2 as network_dot_grpc_dot_fedn__pb2
 
-GRPC_GENERATED_VERSION = '1.69.0'
+GRPC_GENERATED_VERSION = '1.68.1'
 GRPC_VERSION = grpc.__version__
 _version_not_supported = False
